+++ vpin-studio-rest-client/src/main/java/de/mephisto/vpin/restclient/ValidationCode.java
@@ -6,6 +6,7 @@ public interface ValidationCode {
   int CODE_VPX_NOT_EXISTS = 4;
 
   int CODE_NO_DIRECTB2S_OR_PUPPACK = 20;
+  int CODE_NO_DIRECTB2S_AND_PUPPACK_DISABLED = 21;
 
   int CODE_NO_AUDIO = 30;
   int CODE_NO_AUDIO_LAUNCH = 31;