+++ vpin-studio-server/src/main/java/de/mephisto/vpin/server/puppack/PupPack.java
@@ -93,6 +93,10 @@ public class PupPack {
     return true;
   }
 
+  public String getName() {
+    return this.packFolder.getName();
+  }
+
   public File getPupPackFolder() {
     return this.packFolder;
   }