+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/tables/TablesSidebarPUPPackController.java
@@ -9,6 +9,7 @@ import de.mephisto.vpin.restclient.representations.GameRepresentation;
 import de.mephisto.vpin.restclient.representations.PupPackRepresentation;
 import de.mephisto.vpin.restclient.representations.ValidationState;
 import de.mephisto.vpin.ui.Studio;
+import de.mephisto.vpin.ui.events.EventManager;
 import de.mephisto.vpin.ui.tables.validation.LocalizedValidation;
 import de.mephisto.vpin.ui.tables.validation.ValidationTexts;
 import de.mephisto.vpin.ui.util.Dialogs;
@@ -146,6 +147,7 @@ public class TablesSidebarPUPPackController implements Initializable {
     if (game.isPresent() && game.get().isPupPackAvailable()) {
       GameRepresentation g = game.get();
       Studio.client.getPupPackService().setPupPackEnabled(g.getId(), enabledCheckbox.isSelected());
+      EventManager.getInstance().notifyTableChange(this.game.get().getId());
     }
   }